Issue #369 Remove old floating bookmark component and clean up
