+++ src/components/main/header_component.tsx
@@ -8,7 +8,7 @@ import { Id as TaskId, RemoveFromSavedListAction, AddToSavedListAction } from '.
 import { BackButtonComponent } from '../header_button/back_button_component';
 import { HelpButtonComponent } from '../header_button/help_button_component';
 import { MenuButtonComponent } from '../header_button/menu_button_component';
-import { BookmarkButtonComponent } from '../header_button/bookmark_button_component';
+import { BookmarkButtonComponent, BookmarkButtonDisplay } from '../bookmark_button/bookmark_button_component';
 import { Locale } from '../../locale';
 import {
     Routes, isOnParentScreen, isOnChildScreen, pathMatchesRoute, getMatchParamsFromPathandRoute,
@@ -65,14 +65,13 @@ const TopicDetailScreenHeader = (props: Props): JSX.Element => {
     const leftButton = <BackButtonComponent history={props.history} />;
     const rightButtons: ReadonlyArray<JSX.Element> = [
         <BookmarkButtonComponent
-            key={0}
             isBookmarked={R.contains(taskId, props.savedTasksIdList)}
             addBookmark={(): AddToSavedListAction => props.addBookmark(taskId)}
             removeBookmark={(): RemoveFromSavedListAction => props.removeBookmark(taskId)}
             textColor={textColor}
+            display={BookmarkButtonDisplay.Default}
         />,
         <MenuButtonComponent
-            key={1}
             onPress={props.onHeaderMenuButtonPress}
             locale={props.currentLocale}
             textColor={textColor}
